requirements.txt
python-dotenv
streamlit
langchain
langchain_openai
gtts